optionalCharacteristics and getCharacteristic();  getService()
As discussed in #121,
@nfarina :
>Perhaps  getCharacteristic  (and  getService ) should be throwing
Errors anyway if the item doesn't >exist instead of returning null,
since they're used often for chaining calls. That way we could >enhance
the description of the Error when you ask for an optional
Characteristic, suggesting to add it first?

@volschin :
>I reopen this. Using addCharacteristic is the same way it is used in
HomeKitTypes.js for required characteristics. You have no added value,
no consistency check, that your added characteristic is allowed in this
context.

>I think there should be a user function for those optional
characteristics, that check this is an allowed one with the data from
this.optionalCharacteristics = []; .

@snowdd1
If  Service.prototype.getCharacteristic(Characteristic.Type)  does not
find the characteristic, AND it is no optionalCharacteristic match, it
fails ungracefully and throws an error; programmer can use
addCharacteristic()  to add it anyhow, but will then receive a WARNING
in console log entry.
If  Service.prototype.getCharacteristic(Characteristic.Type)  does not
find the characteristic, BUT the type is in optionalCharacteristic, it
adds the characteristic.type and returns it.

If  Accessory.prototype.getService(Service.type)  does not find the
service in the accessory, it fails ungracefully with an exception.