Missing Characteristic
Line 2954 of the Service.Faucet has been edited to add the "InUse" characteristic. Without this, the  iOS Home app can't tell when the faucet has turned on or off.
I realize the file says that HomeKitTypes.js is auto-generated and should not be manually edited, but was unclear how best to flag this error. I assume you'll be able to fix it in the proper place and re-generate this file.